feat: organize components - components now organized by the page they are displayed on - cards and lists directories removed - notes on the page showing which components belong removed - reorganization working
