+++ src/components/forms/PipelineSetUp.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 import { SubmitHandler, useForm } from "react-hook-form";
-import { pipelineFormSchema, PipelineFormType } from "../schema/formSchema";
+import { pipelineFormSchema, PipelineFormType } from "../../schema/formSchema";
 import { zodResolver } from "@hookform/resolvers/zod";
 import axios from "axios";
