+++ src/components/forms/ServiceSetUp.tsx
@@ -1,11 +1,12 @@
-import { useEffect } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import { useState, useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 import { SubmitHandler, useForm } from "react-hook-form";
-import { serviceEditFormSchema, ServiceEditFormType } from "../schema/formSchema";
+import { serviceFormSchema, ServiceFormType } from "../../schema/formSchema";
 import { zodResolver } from "@hookform/resolvers/zod";
 import axios from "axios";
 
 const TEST_SERVICES_URL = import.meta.env.VITE_TEST_SERVICES_URL;
+const TEST_PIPELINES_URL = import.meta.env.VITE_TEST_PIPELINES_URL;
 
 const submitButtonStyle = "mt-10 mb-10 bg-transparent hover:bg-indigo-500 text-indigo-700 font-semibold hover:text-white py-2 px-4 border border-indigo-500 hover:border-transparent rounded";
 
@@ -13,38 +14,22 @@ const errorMsgStyle = "bg-red-100 px-4 py-2 text-red-700";
 
 const inputBorderStyle = "ml-6 border border-gray-300 rounded-md py-2 px-4 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500";
 
-const editableFields = [
-  'name',
-  'triggerOnMain',
-  'triggerOnPrOpen',
-  'triggerOnPrSync',
-  'useStaging',
-  'autoDeploy',
-  'githubRepoUrl',
-  'unitTestCommand',
-  'integrationTestCommand',
-  'codeQualityCommand',
-  'dockerfilePath',
-  'dockerComposeFilePath',
-]
-
-const ServiceEdit = () => {
+const ServiceSetUp = () => {
+  const [pipelineId, setPipelineId] = useState("");
   const navigate = useNavigate();
-  const { serviceId } = useParams();
 
   const {
     register,
     handleSubmit,
-    setValue,
     formState: { errors },
-  } = useForm<ServiceEditFormType>({
-    resolver: zodResolver(serviceEditFormSchema),
-});
+  } = useForm<ServiceFormType>({
+    resolver: zodResolver(serviceFormSchema),
+  });
 
-  const onSubmit: SubmitHandler<ServiceEditFormType> = async (editedData) => {
+  const onSubmit: SubmitHandler<ServiceFormType> = async (data) => {
+    data.pipelineId = pipelineId;
     try {
-      await axios.patch(TEST_SERVICES_URL + serviceId, editedData);
-      alert('Service is being upated.')
+      await axios.post(TEST_SERVICES_URL, data);
       navigate('/services');
     } catch (e) {
       console.log(e);
@@ -52,24 +37,17 @@ const ServiceEdit = () => {
   }
 
   useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await axios.get(TEST_SERVICES_URL + serviceId);
-  
-        editableFields.forEach(field => {
-          setValue(field, response.data[field])
-        })
-      } catch (e) {
-        console.log(e);
-      }
+    const fetchPipeline = async () => {
+      const response = await axios.get(TEST_PIPELINES_URL);
+      setPipelineId(response.data[0].id);
     }
-    fetchData();
+    fetchPipeline();
   }, []);
-
+  
   return (
     <div className="mt-8 ml-8">
     <h2 className="text-3xl text-indigo-700 font-extrabold mb-4">Service Set Up</h2>
-    <p className="mb-4">Pipeline ID: (retrieve from backend)</p>
+    <p className="mb-4">Pipeline ID: {pipelineId}</p>
     <form onSubmit={handleSubmit(onSubmit)}>
       <div className = "flex flex-col gap-2 w-64">
 
@@ -231,14 +209,48 @@ const ServiceEdit = () => {
           )}
         </div>
 
+
+        <h3 className="text-xl text-indigo-700 font-bold mb-4 mt-8">AWS ECR Repo</h3>
+        <div className = "flex flex-col gap-2 w-64">
+          <label htmlFor='awsEcrRepo'>AWS ECR Repo: </label>
+          <input 
+            className={inputBorderStyle}
+            type="text"
+            id="awsEcrRepo"
+            {...register("awsEcrRepo")}  
+          />
+          {errors.awsEcrRepo && (
+            <span className={errorMsgStyle}>
+              {errors.awsEcrRepo?.message}
+            </span>
+          )}
+        </div>
+
+
+        <h3 className="text-xl text-indigo-700 font-bold mb-4 mt-8">AWS ECS Service</h3>
+        <div className = "flex flex-col gap-2 w-64">
+          <label htmlFor='awsEcsService'>AWS ECS Service: </label>
+          <input 
+            className={inputBorderStyle}
+            type="text"
+            id="awsEcsService"
+            {...register("awsEcsService")}  
+          />
+          {errors.awsEcsService && (
+            <span className={errorMsgStyle}>
+              {errors.awsEcsService?.message}
+            </span>
+          )}
+        </div>
+
         </div>
       </div>
       <button 
       className={submitButtonStyle}
-      type="submit">Update Service</button>
+      type="submit">Continue To View Services</button>
     </form>
   </div>
   )
 }
 
-export default ServiceEdit;
+export default ServiceSetUp
