+++ src/components/run/Run.tsx
@@ -1,10 +1,10 @@
 import { useParams } from "react-router";
 import { useState, useEffect } from 'react';
 import axios from "axios";
-import { RunType } from "../schema/runSchema";
-import StagesList from "./lists/StagesList";
-import RunHeaderCard from "./cards/RunHeaderCard";
-import { StageType } from "../schema/stageSchema";
+import { RunType } from "../../schema/runSchema";
+import StagesList from "./StagesList";
+import RunHeaderCard from "./RunHeaderCard";
+import { StageType } from "../../schema/stageSchema";
 
 
 const TEST_STAGES_URL = import.meta.env.VITE_TEST_STAGES_URL;
@@ -51,11 +51,7 @@ const Run = () => {
   
   return (
     <div className="mt-8 ml-8">
-    <h2 className="text-3xl text-indigo-700 font-extrabold mb-4">Run Stages</h2>
-
-    {/* DISPLAY INFO TO REMOVE */}
-    <h2 className="text-1xl text-indigo-700 font-bold mb-4">Run.tsx - RunHeaderCard.tsx - StagesList.tsx - StageCard.tsx</h2>
-    <h2 className="text-1xl text-indigo-700 font-bold mb-4">/runs/:runId</h2>
+    <h2 className="text-3xl text-indigo-700 font-extrabold mb-4">Run</h2>
 
     <RunHeaderCard run={run} />
     <div className="border rounded-lg shadow-md p-4 mr-80">
