+++ src/components/services/Services.tsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
-import { ServiceType } from "../schema/serviceSchema";
-import ServicesList from "./lists/ServicesList";
-import ServicesHeaderCard from "./cards/ServicesHeaderCard";
-import { PipelineType } from "../schema/pipelineSchema";
+import { ServiceType } from "../../schema/serviceSchema";
+import ServicesList from "./ServicesList";
+import ServicesHeaderCard from "./ServicesHeaderCard";
+import { PipelineType } from "../../schema/pipelineSchema";
 import axios from 'axios';
 
 const defaultPipeline = {
@@ -44,10 +44,6 @@ const Services = () => {
   return ( 
     <div className="mt-8 ml-8">
     <h2 className="text-3xl text-indigo-700 font-extrabold mb-4">Services</h2>
-    
-    {/* DISPLAY INFO TO REMOVE */}
-    <h2 className="text-1xl text-indigo-700 font-bold mb-4">Services.tsx - ServicesHeaderCard.tsx - ServicesList.tsx - ServiceCard.tsx</h2>
-    <h2 className="text-1xl text-indigo-700 font-bold mb-4">/services</h2>
     
     <ServicesHeaderCard pipeline={pipeline} />
     <div className="border rounded-lg shadow-md p-4 mr-80">
